feat: user can edit birthday
The plan was to have a hidden date input field whose value would be bound to the dropdown values of day, month, year using Alpine. However we ran into several issues with this approach since we also needed this date input field to be bound to the $user model's dob field. It proved to be a lot easier to define 3 properties ($day, $month, $year) in our Livewire component and then calculate $user.dob server-side.